created a builder for the airport config that will eventually translate an api config into a graphic config.

+++ src/com/bnotions/airspacecontrol/entity/AirportConfig.java
@@ -2,10 +2,10 @@ package com.bnotions.airspacecontrol.entity;
 
 public class AirportConfig {
 
-    public double[] center;
-    public double[][] apron;
-    public double[] tower;
-    public double[][] terminals;
+    public int[] center;
+    public int[][] apron;
+    public int[] tower;
+    public int[][] terminals;
     public RouteConfig[] runways;
     public RouteConfig[] taxiways;
